Created portal tiles and portal controls

+++ Playerdom.Shared/Chunk.cs
@@ -8,7 +8,7 @@ namespace Playerdom.Shared
     [MessagePackObject]
     public struct Chunk
     {
-        public const uint SIZE = 32;
+        public const byte SIZE = 32;
 
         [Key(0)]
         public Tile[,] Tiles
